templates/requirements.txt
requests==2.32.5
pyyaml==6.0.3
python-dotenv==1.2.2
